Added the time feature in digital_time_capsule.py
when the user clicks on the button 'create memory' the time is stored into several different variables. For the reminders and features like that I created a total_days variable so that it will be easier to compute different reminders and add them to the queue, but I have still kept the normal full date in different variables like final_month and final_day.

I have also experimented with a new colour theme that could work, but that is still a work in progress and I want to finish the main code logic without worrying too much about the aesthetics.