Update digital_time_capsule.py
I added the home screen, I added a lot of quality of life features like being able to delete entries and added the reminders entry although it currently looks horrible. My next order of business is to fix up whatever bugs there currently are, then display the reminder date on the app, then find a way to actually notify the user and display the message in the home screen if the due date has already passed and then afterwards I need to add an edit button in the memory log where it redirects the user into the create memory area and allows the user to edit the memory.